Reorganize the tests into tracing an API files.
